No longer need an inner class to delay loading. LootTable is not possible.
Since we already initialize things earlier, the inner class is no longer
required to properly initialize this array. So it is now a part of
WorldGenGarden now.

Loot Table is nice, except it is designed for Item. There is no reliable
way to look up Block based on Item. So just give up on the idea of
moving to loot table.