Non-TEISR model for chopping board. May increase performance because there is no longer direct GL for item.
Old codes are kept in case things are screwed. When it's ready, old codes
shall be removed.

Credits to tterrag (the one who points out the correct way to implement
this, also generously provides necessary data) and quadraxis (the one
who opened a PR that will potentially simplify that messy ImmutableMap)