WTF?
Forge docs (as of Jan 28th 2019 UTC-8) claims that cap. data are synced
if it's returning from the End. I was told that the data was actually
lost after returning from the End with current setup. I don't know
which is the case now. It's not expensive to copy data either... so
here we go... unconditionally copy the data...